+++ resources/lang/ro/response.php
@@ -15,5 +15,6 @@ return [
 		'city' => 'oraș|orase',
 		'timezone' => 'fus orar|fusuri orare',
 		'currency' => 'valută|valute',
+		'language' => 'limba|limbi',
 	],
 ];