+++ src/Http/Controllers/Language/LanguageController.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Nnjeim\World\Http\Controllers\Language;
+
+use Nnjeim\World\Http\Controllers\BaseController;
+
+class LanguageController extends BaseController
+{
+	protected string $requestBasePath = 'Nnjeim\\World\\Http\\Requests\\Language';
+
+	protected string $actionBasePath = 'Nnjeim\\World\\Actions\\Language';
+}