+++ src/Http/Requests/Language/IndexRequest.php
@@ -1,0 +1,21 @@
+<?php
+
+namespace Nnjeim\World\Http\Requests\Language;
+
+use Nnjeim\World\Http\Requests\BaseRequest;
+
+class IndexRequest extends BaseRequest
+{
+	/**
+	 * Get the validation rules that apply to the request.
+	 *
+	 * @return array
+	 */
+	public function rules(): array
+	{
+		return [
+			'fields' => 'sometimes',
+			'filters' => 'sometimes',
+		];
+	}
+}