+++ src/Routes/index.php
@@ -12,48 +12,20 @@ Route::group([
 		'bindings',
 	],
 ], function () {
-	/*-- Countries ---------------------------------------------------------------------------------------------------*/
-	Route::group([
-		'prefix' => 'countries',
-		'as' => 'countries',
-	], function () {
-		Route::get('/', [Controllers\Country\CountryController::class, 'index'])->name('index');
-	});
-	/*-- States ------------------------------------------------------------------------------------------------------*/
-	Route::group([
-		'prefix' => 'states',
-		'as' => 'states',
-	], function () {
-		Route::get('/', [Controllers\State\StateController::class, 'index'])->name('index');
-	});
-	/*-- Cities ------------------------------------------------------------------------------------------------------*/
-	Route::group([
-		'prefix' => 'cities',
-		'as' => 'cities',
-	], function () {
-		Route::get('/', [Controllers\City\CityController::class, 'index'])->name('index');
-	});
-	/*-- Timezones ---------------------------------------------------------------------------------------------------*/
-	Route::group([
-		'prefix' => 'timezones',
-		'as' => 'timezones',
-	], function () {
-		Route::get('/', [Controllers\Timezone\TimezoneController::class, 'index'])->name('index');
-	});
-	/*-- Currencies --------------------------------------------------------------------------------------------------*/
-	Route::group([
-		'prefix' => 'currencies',
-		'as' => 'currencies',
-	], function () {
-		Route::get('/', [Controllers\Currency\CurrencyController::class, 'index'])->name('index');
-	});
-	/*-- Phones ------------------------------------------------------------------------------------------------------*/
-	Route::group([
-		'prefix' => 'phones',
-		'as' => 'phones',
-	], function () {
-		Route::post('/validate', [Controllers\Phone\PhoneController::class, 'validate'])->name('validate');
-		Route::post('/format', [Controllers\Phone\PhoneController::class, 'format'])->name('format');
-		Route::post('/strip', [Controllers\Phone\PhoneController::class, 'strip'])->name('strip');
-	});
+	// countries
+	Route::get('/countries', [Controllers\Country\CountryController::class, 'index'])->name('countries.index');
+	// states
+	Route::get('/states', [Controllers\State\StateController::class, 'index'])->name('states.index');
+	// cities
+	Route::get('/cities', [Controllers\City\CityController::class, 'index'])->name('cities.index');
+	// timezones
+	Route::get('/timezones', [Controllers\Timezone\TimezoneController::class, 'index'])->name('timezones.index');
+	// currencies
+	Route::get('/currencies', [Controllers\Currency\CurrencyController::class, 'index'])->name('currencies.index');
+	// languages
+	Route::get('/languages', [Controllers\Language\LanguageController::class, 'index'])->name('languages.index');
+	// phones
+	Route::post('/phones/validate', [Controllers\Phone\PhoneController::class, 'validate'])->name('phones.validate');
+	Route::post('/phones/format', [Controllers\Phone\PhoneController::class, 'format'])->name('phones.format');
+	Route::post('/phones/strip', [Controllers\Phone\PhoneController::class, 'strip'])->name('phones.strip');
 });